gptstudio: Use Large Language Models Directly in your Development Environment

Large language models are readily accessible via API. This package lowers the barrier to use the API inside of your development environment. For more on the API, see https://platform.openai.com/docs/introduction.

See also

Author

Maintainer: James Wade github@jameshwade.com (ORCID) [copyright holder]

Authors:

Michel Nivard m.g.nivard@vu.nl [copyright holder]

Samuel Calderon samuel.calderon@uarm.pe (ORCID)