Current Configuration for gptstudio

This function prints out the current configuration settings for gptstudio and checks API connections if verbose is TRUE.

Usage

gptstudio_sitrep(verbose = TRUE)

Arguments

verbose

Logical value indicating whether to output additional information, such as API connection checks. Defaults to TRUE.

Value

Invisibly returns NULL, as the primary purpose of this function is to print to the console.

Examples

gptstudio_sitrep(verbose = FALSE)  # Print basic settings, no API checks
#> 
#> ── Configuration for gptstudio ─────────────────────────────────────────────────
#> No user configuration file found at
#> /home/runner/.config/R/gptstudio/config.yml. Using default configuration.
#> Change configuration settings in the chat app. Lauch the chat app with addins
#> or `gptstudio_chat()`.
#> 
#> ── Current Settings ──
#> 
#> - Model: gpt-3.5-turbo
#> - Task: coding
#> - Language: en
#> - Service: openai
#> - Custom prompt: You are a helpful assistant.
#> - Stream: TRUE
#> - Code style: no preference
#> - Skill: beginner
#> Run `gptstudio_sitrep(verbose = TRUE)` to check API connections.
#> ── End of gptstudio configuration ──────────────────────────────────────────────
gptstudio_sitrep()                 # Print settings and check API connections
#> 
#> ── Configuration for gptstudio ─────────────────────────────────────────────────
#> No user configuration file found at
#> /home/runner/.config/R/gptstudio/config.yml. Using default configuration.
#> Change configuration settings in the chat app. Lauch the chat app with addins
#> or `gptstudio_chat()`.
#> 
#> ── Current Settings ──
#> 
#> - Model: gpt-3.5-turbo
#> - Task: coding
#> - Language: en
#> - Service: openai
#> - Custom prompt: You are a helpful assistant.
#> - Stream: TRUE
#> - Code style: no preference
#> - Skill: beginner
#> 
#> ── Checking API connections ──
#> 
#> ── Checking OpenAI API connection 
#> ✖ API key is not set or invalid for OpenAI service.
#> ✖ Failed to connect to the OpenAI API service.
#> 
#> ── Checking HuggingFace API connection 
#> ✖ API key is not set or invalid for HuggingFace service.
#> ✖ Failed to connect to the HuggingFace API service.
#> 
#> ── Checking Anthropic API connection 
#> ✖ API key is not set or invalid for Anthropic service.
#> ✖ Failed to connect to the Anthropic API service.
#> 
#> ── Checking Google PALM Makersuite API connection 
#> ✖ API key is not set or invalid for Google PALM service.
#> ✖ Failed to connect to the Google PALM API service.
#> 
#> ── Checking Azure OpenAI API connection 
#> ✖ API key is not set or invalid for Azure OpenAI service.
#> ✖ Failed to connect to the Azure OpenAI API service.
#> 
#> ── Checking Perplexity API connection 
#> ✖ API key is not set or invalid for Perplexity service.
#> ✖ Failed to connect to the Perplexity API service.
#> 
#> ── Check Ollama for Local API connection 
#> ✖ Couldn't connect to Ollama in <http://localhost:11434>. Is it running there?
#> 
#> ── Getting help ──
#> 
#> See the gptstudio homepage (<https://michelnivard.github.io/gptstudio/>) for
#> getting started guides and package documentation. File an issue or contribute
#> to the package at the GitHub repo
#> (<https://github.com/MichelNivard/gptstudio>).
#> ── End of gptstudio configuration ──────────────────────────────────────────────